bug fixing

+++ CodeTools/OtherTools/RandomTools.cs
@@ -14,7 +14,7 @@ namespace UniTools
             if (chance == 0) return;
             if (GetChance(chance)) action?.Invoke();
         }
-        public static bool GetChance(int chance) => UnityEngine.Random.Range(0, 100) <= chance;
+        public static bool GetChance(float chance) => UnityEngine.Random.Range(0, 100) <= chance;
         public static Vector3 GetRandomPointInRange(Vector3 centre, float range, bool useX = true, bool useY = true, bool useZ = true)
         {
             return new Vector3(centre.x + (useX ? GetRandom() : 0), centre.y + (useY ? GetRandom() : 0), centre.z + (useZ ? GetRandom() : 0));
